refactor(wss): fold batch-of-one helpers into BatchWorld, drop the vestigial batch_execute adapter
After the multi-op trials were extracted, the batch_execute adapter held
only run_batch_of_one + blob_token — BatchWorld's single-op invocation
helpers. Move them onto BatchWorld (run_batch_of_one takes &self; blob_token
is an associated fn) and delete the adapter. The 6 single-path batch invokers
now call w.run_batch_of_one(..) / BatchWorld::blob_token(..). The deleted
module's out-of-scope note (multi-op transaction-integrity lives in
turbovault-nbl.17, not WSS) is folded into BatchWorld's doc comment, and the
adapters mod doc drops the stale batch_execute self-builds-trials clause
(batch rides single_path_trials).

Pure code move — the trial set is unchanged (wss_matrix: 520 passed / 0
failed / 193 ignored).
