+++ crates/turbovault/tests/write_safety_suite/adapters/create_from_template.rs
@@ -10,7 +10,6 @@
 
 use std::collections::HashMap;
 
-use super::batch_execute::run_batch_of_one;
 use super::{Case, SinglePathOp};
 use crate::harness::backend::{Backend, BatchWorld, Layer, MSG, ToolsWorld, observe};
 use crate::harness::outcome::{Observed, Outcome as O};
@@ -95,7 +94,7 @@ impl SinglePathOp<BatchWorld> for CreateFromTemplate {
             fields: fields(),
             force,
         };
-        run_batch_of_one(w, op, rel).await
+        w.run_batch_of_one(op, rel).await
     }
 
     fn ok_effect(&self, observed: &Observed) -> Result<(), String> {
