+++ crates/turbovault/tests/write_safety_suite/adapters/delete_note.rs
@@ -6,7 +6,6 @@
 //! tool-layer behavior today, not substrate-layer, so at this layer it's a
 //! deferred one-off (noted below), tracked with the substrate move of oz6.
 
-use super::batch_execute::{blob_token, run_batch_of_one};
 use super::{Case, SinglePathOp};
 use crate::harness::backend::{Backend, BatchWorld, Layer, MSG, ManagerWorld, ToolsWorld, observe};
 use crate::harness::outcome::{Observed, Outcome as O};
@@ -100,10 +99,10 @@ impl SinglePathOp<BatchWorld> for DeleteNote {
     async fn invoke(&self, w: &BatchWorld, rel: &str, pc: Precondition) -> Observed {
         let op = BatchOperation::DeleteNote {
             path: rel.to_string(),
-            expected_hash: blob_token(&pc),
+            expected_hash: BatchWorld::blob_token(&pc),
             on_backlinks: None,
         };
-        run_batch_of_one(w, op, rel).await
+        w.run_batch_of_one(op, rel).await
     }
 
     fn ok_effect(&self, observed: &Observed) -> Result<(), String> {
