+++ crates/turbovault/tests/write_safety_suite/adapters/manage_tags.rs
@@ -7,7 +7,6 @@
 //! (cutover: qae.9.1). The dirty-gate / precond-vs-workdir cells are `pending`
 //! (nbl.8 burndown).
 
-use super::batch_execute::{blob_token, run_batch_of_one};
 use super::{Case, SinglePathOp};
 use crate::harness::backend::{Backend, BatchWorld, Layer, MSG, ToolsWorld, observe};
 use crate::harness::outcome::{Observed, Outcome as O};
@@ -77,9 +76,9 @@ impl SinglePathOp<BatchWorld> for ManageTags {
             path: rel.to_string(),
             operation: "add".to_string(),
             tags: vec![TAG.to_string()],
-            expected_hash: blob_token(&pc),
+            expected_hash: BatchWorld::blob_token(&pc),
         };
-        run_batch_of_one(w, op, rel).await
+        w.run_batch_of_one(op, rel).await
     }
 
     fn ok_effect(&self, observed: &Observed) -> Result<(), String> {
