+++ crates/turbovault/tests/write_safety_suite/adapters/write_note.rs
@@ -5,7 +5,6 @@
 //! passing the [`Precondition`] directly. The tool layer does not take a
 //! precondition yet, so this does not compile until the cutover (qae.9.1).
 
-use super::batch_execute::run_batch_of_one;
 use super::{Case, SinglePathOp};
 use crate::harness::backend::{Backend, BatchWorld, Layer, MSG, ManagerWorld, ToolsWorld, observe};
 use crate::harness::outcome::{Observed, Outcome as O};
@@ -116,7 +115,7 @@ impl SinglePathOp<BatchWorld> for WriteNote {
                 expected_hash: None,
             },
         };
-        run_batch_of_one(w, op, rel).await
+        w.run_batch_of_one(op, rel).await
     }
 
     fn ok_effect(&self, observed: &Observed) -> Result<(), String> {
